Refactor : Model 폴더 추가 및 User와 Question 클래스를 Model 폴더로 분리
