+++ src/main/java/com/codessquad/qna/repository/QuestionRepository.java
@@ -1,5 +1,6 @@
 package com.codessquad.qna.repository;
 
+import com.codessquad.qna.model.Question;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
